added delete for vessel owner contacts

+++ src/Controller/VesselOwnerContactsController.php
@@ -45,7 +45,7 @@ class VesselOwnerContactsController extends AppController {
  */
 	public function view($id = null) {
 		if($this->request->params['_ext']){
-			$vesselOwnerContact = $this->VesselOwnerContacts->get($id, [
+			$ = $this->VesselOwnerContacts->get($id, [
 			//	'fields' => ['id','name','category_id', 'city_id']
 			]);
 			$this->set('vesselOwnerContact', $vesselOwnerContact);
@@ -60,53 +60,6 @@ class VesselOwnerContactsController extends AppController {
 	}
 
 /**
- * Add method
- *
- * @return void
- */
-	public function add() {
-		$vesselOwnerContact = $this->VesselOwnerContacts->newEntity($this->request->data);
-		if ($this->request->is('post')) {
-			if ($this->VesselOwnerContacts->save($vesselOwnerContact)) {
-				$this->Flash->success('The vessel owner contact has been saved.');
-				return $this->redirect(['action' => 'index']);
-			} else {
-				$this->Flash->error('The vessel owner contact could not be saved. Please, try again.');
-			}
-		}
-		$creators = $this->VesselOwnerContacts->Creators->find('list');
-		$modifiers = $this->VesselOwnerContacts->Modifiers->find('list');
-		$vesselOwners = $this->VesselOwnerContacts->VesselOwners->find('list');
-		$this->set(compact('vesselOwnerContact', 'creators', 'modifiers', 'vesselOwners'));
-	}
-
-/**
- * Edit method
- *
- * @param string $id
- * @return void
- * @throws \Cake\Network\Exception\NotFoundException
- */
-	public function edit($id = null) {
-		$vesselOwnerContact = $this->VesselOwnerContacts->get($id, [
-			'contain' => []
-		]);
-		if ($this->request->is(['patch', 'post', 'put'])) {
-			$vesselOwnerContact = $this->VesselOwnerContacts->patchEntity($vesselOwnerContact, $this->request->data);
-			if ($this->VesselOwnerContacts->save($vesselOwnerContact)) {
-				$this->Flash->success('The vessel owner contact has been saved.');
-				return $this->redirect(['action' => 'index']);
-			} else {
-				$this->Flash->error('The vessel owner contact could not be saved. Please, try again.');
-			}
-		}
-		$creators = $this->VesselOwnerContacts->Creators->find('list');
-		$modifiers = $this->VesselOwnerContacts->Modifiers->find('list');
-		$vesselOwners = $this->VesselOwnerContacts->VesselOwners->find('list');
-		$this->set(compact('vesselOwnerContact', 'creators', 'modifiers', 'vesselOwners'));
-	}
-
-/**
  * Delete method
  *
  * @param string $id
@@ -114,13 +67,26 @@ class VesselOwnerContactsController extends AppController {
  * @throws \Cake\Network\Exception\NotFoundException
  */
 	public function delete($id = null) {
-		$vesselOwnerContact = $this->VesselOwnerContacts->get($id);
-		$this->request->allowMethod(['post', 'delete']);
-		if ($this->VesselOwnerContacts->delete($vesselOwnerContact)) {
-			$this->Flash->success('The vessel owner contact has been deleted.');
-		} else {
-			$this->Flash->error('The vessel owner contact could not be deleted. Please, try again.');
+		if($this->request->params['_ext']){
+	        $vesselOwnerContact = $this->VesselOwnerContacts->get($id);
+	        $message = 'Deleted';
+	        if (!$this->VesselOwnerContacts->delete($vesselOwnerContact)) {
+	            $message = 'Error';
+	        }
+	        $this->set([
+	            'message' => $message,
+	            '_serialize' => ['message']
+	        ]);
 		}
-		return $this->redirect(['action' => 'index']);
+		else {		
+			$vesselOwnerContact = $this->VesselOwnerContacts->get($id);
+			$this->request->allowMethod(['post', 'delete']);
+			if ($this->VesselOwnerContacts->delete($vesselOwnerContact)) {
+				$this->Flash->success('The vesselOwnerContact has been deleted.');
+			} else {
+				$this->Flash->error('The vesselOwnerContact could not be deleted. Please, try again.');
+			}
+			return $this->redirect(['action' => 'index']);
+		}
 	}
 }
